FPS Independant Movement Calculation
- Fixed compilation on debian 8
- Switched position of text from uint to float to store 1/2 pixels

+++ object.h
@@ -39,8 +39,8 @@ typedef struct object_base_st_ {
     llist_elem_st elem;
     object_type_et obj_type;
     obj_api_st *obj_api;
-    uint32_t x;
-    uint32_t y;
+    float x;
+    float y;
     uint32_t flags;
 } object_base_st;
